+++ lib/blazer/detect_anomalies.R
@@ -1,0 +1,14 @@
+tryCatch({
+  library(AnomalyDetection)
+
+  args <- commandArgs(trailingOnly = TRUE)
+
+  con <- textConnection(args[1])
+  data <- read.csv(con, stringsAsFactors = FALSE)
+  data$timestamp <- as.POSIXct(data$timestamp)
+
+  res = AnomalyDetectionTs(data, direction = "both", alpha = 0.05)
+  write.csv(res$anoms)
+}, error = function (e) {
+  write.csv(geterrmessage())
+})